+++ go-api/cmd/mock-api/main.go
@@ -17,7 +17,8 @@ func main() {
 	// start
 	port := "8090"
 	fmt.Println("app is listening and serving on port", port)
-	if err := http.ListenAndServe(fmt.Sprintf(":%s", port), nil); err != nil {
+	err := http.ListenAndServe(fmt.Sprintf(":%s", port), nil)
+	if err != nil {
 		panic(err)
 	}
 }
