+++ go-api/internal/cache/cache.go
@@ -1,0 +1,7 @@
+package cache
+
+import "os"
+
+func IsEnabled() bool {
+	return os.Getenv("MONGODB_ENABLED") == "true"
+}
